More fixes.
Templates now look pretty good and will load eachother. Figured out how
to click hijack in the iframe. Need to grab click sender and search for
"aro" attributes and present a modal to allow editing of content.

Then need to figure out how to sync this back to the server of course,
but that should be too bad... Right?????